+++ lib/ui/explore/ExplorePanel.dart
@@ -1106,8 +1106,8 @@ class ExplorePanelState extends State<ExplorePanel>
                   itemCount: filterValues.length,
                   itemBuilder: (context, index) {
                     return FilterListItemWidget(
-                      label: filterValues[index],
-                      subLabel: hasSubLabels ? filterSubLabels![index] : null,
+                      title: filterValues[index],
+                      description: hasSubLabels ? filterSubLabels![index] : null,
                       selected: (selectedFilter?.selectedIndexes != null && selectedFilter!.selectedIndexes.contains(index)),
                       onTap: () {
                         Analytics().logSelect(target: "FilterItem: ${filterValues[index]}");
@@ -1147,10 +1147,9 @@ class ExplorePanelState extends State<ExplorePanel>
       int filterValueIndex = selectedFilter.firstSelectedIndex;
       String? filterHeaderLabel = filterValues[filterValueIndex];
       filterTypeWidgets.add(FilterSelectorWidget(
-        label: filterHeaderLabel,
+        title: filterHeaderLabel,
         hint: _getFilterHintByType(selectedFilter.type),
         active: selectedFilter.active,
-        visible: true,
         onTap: (){
           Analytics().logSelect(target: "Filter: $filterHeaderLabel");
           return _onFilterTypeClicked(selectedFilter);},